requirements.txt
# torch==2.0.0
torchsummary==1.5.1
torchvision==0.15.1
Pillow==9.4.0
tqdm==4.64.1
pandas==1.4.4
